Some javadoc
